[refactor] Study Join 패키지 변경
