+++ app/src/main/java/com/ssu/better/presentation/ui/study/join/StudyJoinScreen.kt
@@ -1,4 +1,4 @@
-package com.ssu.better.presentation.ui.study_join
+package com.ssu.better.presentation.ui.study.join
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
